requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
torch>=2.0
matplotlib>=3.7
joblib>=1.3
